day1_a2/requirements.txt
fastapi
uvicorn
groq
python-dotenv
pydantic
aiofiles
python-multipart
